ci: Add workflow for testing action
Closes #10

Adds a workflow to run tests on GitHub Actions Continuous Integration.

It only runs `pull_request` for certain events and even more specifically on changes to specific files (see the `paths`
list under `pull_request`).

At the same time I've linked the `sudlab/IsoSLAM` repository to [codecov.io](https://codecov.io) so that the tests,
which generate code coverage reports (in XML) are uploaded and available for analysis automatically. Required adding a
Secret Token from Codecov to the repository.

May be tweaked in future but this is a good starting point.