tests: Initial setup of tests
Adds a bunch of input files to `tests/resources/*` although many of these are input to the full pipeline it is useful to
have for this minimal reproducible example.

There is I think a need to separate the overall pipeline of processing raw output (undertaken by
`pipeline_slam_3UIs.py`) from the actual key analysis that IsoSLAM (the code in
`isoslam/all_introns_conuts_and_info.py`).