feature: Adds processing module and tests
Closes #66

Adds an `isoslam/processing.py` module which creates an argument parser (with sub-parsers) and associated tests that the
arguments are passed through correctly.

In due course we will add a `isoslam/default_config.yaml` and will add functions to update this configuration with the
command line arguments that users can supply (which I expect to grow as more modules get added).

The minimum Python version has been bumped to `3.10` so that type-hints can be written as `X | Y` which is more
succinct.

Also adds an entry point so that once installed can be run with `isoslam [options] <module> [options]`.

Docstrings are in place along with typehints which pass mypy \o/