feature(isoslam): Adds functions to extract features from reads
Closes #88

Adds functions to extract features from a read both individual `pysam.AlignedSegment` and pairs of reads returned from
the `pysam.fetch()` iterator.

- Includes tests.
- Extracts
  - `start`
  - `end`
  - `length`
  - tag for `XS` if present, otherwise `None`
  - tag for `XT` if present, otherwise `None`
  - zips up the blocks and returns them as `block_start` and `block_end`

Tests use fixtures defined in `tests/conftest.py` and I had trouble trying to moock `pysam.AlignedSegment` as a number
of the attributes I wanted to mock were immutable. I therefore opted to load a `.bam` file and extract some individual
`AlignedSegments` as features and use these in the parameterised tests.

However, confusingly the value I used for `start` in the `fetch()` call did not match the first item returned by the
generator (this can be seen by the disconnect between the `start` parameters in the `fetch()` calls and the names of the
fixtures which reflect the values that are tested and used in the parameter names).

I'm not sure why this is, but one consequence of it is that I'm yet to have extracted any `AlignedSegment` where the
`get_tag("XS") == Aligned` and so in turn these are all `None`, similarly there are no `get_tag("XT")` (the transcript
id?).

Guidance on how to identify and extract these would be very welcome (I just stuck some `print()` statements into the
section of `all_introns_counts_and_info.py` where the `get_tag("XT") calls were made to show the `read1_start` at that
point, very crude as I'm still fumbling my way around these files and their specification!).