feature(utils): Rounds out updating configuration
Main function `update_config()` was already in place, this rounds it out by converting the values associated with the
keys `base_dir` and `output_dir` to be objects of type `Path` so they are easy to use in subsequent processing.

Will have to ensure in each processor function that after the configuration is loaded a call is made to
`utils.update_config()` so that values get updated.