refactor: extract untranslated regions from assigned reads
Extracts the untranslated region coordinates from assigned transcript reads.

Reduces code duplication by abstracting repeated code into a function (`isoslam.extract(utron)`) and implements tests.

The regression test which checks that replacing the code in `all_introns_counts_and_info.py` produces the expected
results when replacing the old code with the newer code fails for one of the files here and I would appreciate insight
as to whether this is correct or not as I don't know if its a mistake in what I've introduced or a problem with the code
as it existed.

The failing test is...

```
tests/test_all_introns_counts_and_info.py::test_main[file 1] PASSED                                                                                      [ 50%]
tests/test_all_introns_counts_and_info.py::test_main[file 2] FAILED                                                                                      [100%]

=========================================== FAILURES ============================================================
________________________________________ test_main[file 2] ______________________________________________________

regression test output differences for tests/test_all_introns_counts_and_info.py::test_main[file 2]:
    (recorded output from tests/_regtest_outputs/test_all_introns_counts_and_info.test_main[file_2])

    >   --- current
    >   +++ expected
    >   @@ -819,13 +819,7 @@
    >    0       482  ENST00000442898  18492  24850   9      -        Ret            0           58       219
    >    0       483  ENST00000442898  18492  24850   9      -        Ret            0           58       219
    >    0       484  ENST00000442898  14940  15080   9      -        Ret            1           54       233
    >   -0       484  ENST00000442898  15149  15908   9      -        Ret            1           54       233
    >   -0       484  ENST00000442898  16061  16717   9      -        Ret            1           54       233
    >    0       484  ENST00000442898  16876  16964   9      -        Ret            1           54       233
    >   -0       484  ENST00000442898  17166  17343   9      -        Ret            1           54       233
    >   -0       484  ENST00000442898  17479  17718   9      -        Ret            1           54       233
    >   -0       484  ENST00000442898  17855  18027   9      -        Ret            1           54       233
    >   -0       484  ENST00000442898  18174  18380   9      -        Ret            1           54       233
    >    0       484  ENST00000442898  18492  24850   9      -        Ret            1           54       233
    >    0       485  ENST00000442898  17479  17718   9      -        Spl            0           62       293
    >    0       485  ENST00000442898  17855  18027   9      -        Spl            0           62       293

---------------------------------------- Captured stderr call ---------------------------------------------------
```

Which is a little hard to read but what it means is that in the existing output for this test (found in
`tests/_regtest_outputs/test_all_introns_counts_and_info.test_main[file_2]`) there are the following entries...

```
❱ grep 484 tests/_regtest_outputs/test_all_introns_counts_and_info.test_main\[file_2\].out | sort
0       484  ENST00000442898  14940  15080   9      -        Ret            1           54       233
0       484  ENST00000442898  16876  16964   9      -        Ret            1           54       233
0       484  ENST00000442898  18492  24850   9      -        Ret            1           54       233
```

In the output when running the refactored code there are six more entries which I've marked with a `*` below...

```
❱ grep 484 /tmp/pytest-of-neil/pytest-current/test_main_file_2_current/test.tsv | sort
484	ENST00000442898	14940	15080	9	-	Ret	1	54	233
484	ENST00000442898	15149	15908	9	-	Ret	1	54	233 *
484	ENST00000442898	16061	16717	9	-	Ret	1	54	233 *
484	ENST00000442898	16876	16964	9	-	Ret	1	54	233
484	ENST00000442898	17166	17343	9	-	Ret	1	54	233 *
484	ENST00000442898	17479	17718	9	-	Ret	1	54	233 *
484	ENST00000442898	17855	18027	9	-	Ret	1	54	233 *
484	ENST00000442898	18174	18380	9	-	Ret	1	54	233 *
484	ENST00000442898	18492	24850	9	-	Ret	1	54	233
```

I've no idea whether these six additional entries should be there or not. They look like they perhaps should be as the
`start` and `end` are within the limits of the original three entries but I don't know if that is means they should be
present or not.