feature(io): converts connection to .bed file to dictionary
I'd missed the fact that the bed file connection needs to be actually converted to a dictionary in order to extract
untranslated regions for assigned reads so updates `load_file()` to use the new `_bed_to_dictionary()` function that
converts the connection to a dictionary.

However, rather than returning a dictionary of tuples I've opted to return a dictionary of dictionaries so that we can
name the elements `chr`, `start`, `end`, `transcript_id` and `bedstrand`. I've also opted _not_ to include the
`transcript_id` within each of the nested dictionaries (which were tuples) because these are the keys to the dictionary
and so if we iterate over this at any point we can still get to `transcript_id`.