refactor: building within_intron and spliced subsets
Closes #91
Closes #92

Extracts the duplicated code that builds subsets of reads that are within introns (or span introns) and within
spliced_3ui and abstracts out to functions with tests. Currently the tests fail and I don't think they should so I need
to work out why and perhaps add some additional fixtures to use here.