tests: Meaningful tests for refactored filtering
Identified instances where meaningful results are obtained when filtering for instances where the start/end are within
introns or there is splicing.

Tests are not fully comprehensive yet as not all logic statements are covered, in particular when a transcript spans a
whole region but I'm currently a bit hazy on whether this will always happen as the `query_length` is always
`150` (`read_length` is longer and was the cause of previous errors).