docs: Adds useful diagram and textual description of the workflow
Doesn't close #136 but I wanted to get started on the documentation and start building this up.  My one sentence
statements are likely way off the mark.

Feel free to expand on this either directly on this branch (`ns-rse/136-document-alignment`) or make a branch from it
and we can then merge it back in.

**NB** We have [markdownlint-cli2](https://github.com/DavidAnson/markdownlint-cli2) implemented as a
[pre-commit](https://pre-commit.com). Could add this to virtual environment to lint code locally before commits
otherwise [pre-commit.ci](https://pre-commit.ci) will run on GitHub and fix what it can but there may be some that it
can't (e.g. it doesn't automatically fix hard-wrapping lines at 120 characters which is very useful for narrowing down
where there are changes).