docs: indent mkdocs options to API pages so they render
Thanks to @tdjames1 for the pointer that these need indenting.

I think I did originally try that, but found that [prettier](https://prettier.io/docs/en/) which is one of the many
[pre-commit](https://pre-commit.com) hooks in place reformatted things and didn't pay sufficient attention.

This commit/pull-request includes a `.prettierignore` file to explicitly prevent that from happening.

Unsurprisingly the pages now look a _lot_ better.

Thanks @tdjames1 for letting me know about that :+1: