refactor: Move logic to skip processing to earlier in processing
Moves a lot of the logic that controls whether processing is undertake to earlier in the loop (thus avoiding some
unnecessary processing if we are to ultimately skip things).

At the same time I've taken the opportunity to (maybe) make the logic easier to understand and updated the comments with
explanations of what is being done.