refactor: Function for writing results
Closes #147

Breaks out the duplicated code that writes results to disk to its own function with tests.

I think longer term we may want to see if there are alternatives to having a file connection open and performing
operations within and writing to disk but I will need to see how this plays out with RAM when working with larger
datasets.

- Ignore `.envrc` (a file @ns-rse is using to leverage [direnvs](https://direnv.net/))
- Tidy up some linting in `all_introns_counts_and_info.py`
- Smattering of `TODO` items collated into `TODO.md` for @ns-rse future reference