feature(validation): Adds module/function w/tests to validate config
Closes #73

Adds a module to validate the configuration _after_ loading and updating with any user specified command line
options. Its useful to check these early in processing and exit if things aren't right.

- Test is included.
- Instructions on adding configuration options and how to ensure they too are validated are included in the
  documentation.

One thing I'm unsure about and would appreciate input on @IanSudberry is the string forms that would be valid for the
chromosome field. I've treated as strings because of `X` and `Y` and the possibility the values are prefixed with
`chr#`.

- Is the possibility of the chromosome indicator being prefixed with `chr` correct or not? (its easy to remove them in
  this PR if that notation is never observed)
- Are the x/y chromosomes _always_ upper or lower case or could there be a mixture?