feature(io|summary): Improves flexibility of loading results
Existing functionality to load output files now works with `.parquet`/`.tsv`/`.csv` files and loads data as
[Polars](https://docs.pola.rs/) dataframes.

Tests have been expanded to use x35 files of each type.

This is a precursor to #114/#115 as the p-value summaries duplicate the summary counts so needed to improve flexibility
of this step to work with the multiple file formats now supported.

Next step is to extract the `hour` and `repetition` from the returned data frames and summarise by these (which is
possible since the `groupby` option to `summar_counts.summary()` allows this :-).