feature(summary): counts, total and percent now calculated
After lots of reading and trying to get my head round what is actually required I've started by revising this function
to produce counts by transcript/read/strand and whether there is one or more conversion or not and then aggregating
counts a second time but ignoring whether there is one or more conversion. These are combined and the percentages can be
calculated.

Next step is identifying transcripts where conversions are only False and deriving dummies for these to be True with the
observed `conversion_count = 0`, `conversion_total = <observed value>` and `conversion_percent = 0`.