feature(summary): derive dataset of one or more conversions
Took me a while but the final step in checking the data wrangling prior to statistical analysis is complete.

The `summary_count` which counts the number of conversions by transcript/start/end/assignment/strand and whether there
are one or more conversions, returning the number, total number at each location (the sum of whether there are or are
not conversions) and percentages is used as there are some instances where there hasn't been a conversion.

These are used to create dummies setting the number and percentage of conversions to zero and combining with the
original data.

A subset of data points where conversions are true is then taken and it is the percentages from these that are used as
input to the subsequent modelling steps.

Also...

- ensures day/hour/replicate are integers
- updates some regression tests in light of this change