Remove columns from summary and associated io functions
Had initially thought that loading a subset of columns when building datasets would save time, probably not a lot that
can be saved really though, whilst `chromosome` column is useful and redundant since `transcript_id` always maps to the
same location its useful for users. Instead it is removed from the internal dataframes that are processed to avoid
an unnecessary level when using `group_by` and added to results after summarisation.