refactor: correct calculation of weighted mean of percents
I stupidly took the mean/median of the percentages ignoring that you need to account for differences in the
denominator. Instead, since we have the counts and the totals at each locus we don't need to weight the percentages by
the totals but can just sum them within the grouping and calculate the percentage across replicates directly.

Updates all the tests in light of this (and makes the `simple` examples more realistic).