feature(summary): calculate weights for modelling
Adds `summary._derive_weight_within_isoform()` which calculates the weight for the mean and normalised percentages.

The weights are based on the `conversion_total` observed at each time point after summation across replications. The
total number of events across all replications/time points _within_ a `Transcript_id` / `Strand` / `Start` / `End` /
`Assignment` is derived and the weight is the proportion of events at that time point out of this total.