docs(workflow): Adding details of data preparation for modelling
Attempting to add worked examples showing the flow of data through the data preparation stages, as much for my own
benefit to ensure that I've understood and the steps require and done them correctly.

I still need to add in removing instances where there is not both a `Return` and `Spliced` instance from the workflow
and also possibly remove instances where there is no conversion at baseline since this results in `NaN` when doing a
division by zero (this could be done by modifying the `join()` within `_merge_average_with_baseline()` which currently
defaults to `inner` but should perhaps be `right`, or flip the order and make it `left`!).