feature(summary): Adds _finder_read_pairs() function
Ultimately only pairs where there is `Assignment` for both `Ret` _and_ `Spl` are the only sites that need testing. We
therefore add a function which finds the intersection of the two and return it as a dataframe.

This can be used to filter down the dataset itself (ideally before all of the averaging etc. is done) to reduce the
number of calculations that need to be undertaken.

It will be part of the pipeline in the `Statistics.__post_init__` method.