feature(summary): general function to get different groupings
Realised I was falling foul of the DRY (Don't Repeat Yourself) rule as I had littered functions with similar statements
to get the different groupings by which operations were to be performed within or joined by.

Have therefore abstracted this out to a separate function which takes a string to indicate the grouping that should be
returned.

Because of how the functions have been written its conceivable that users may wish to perform their own level of
grouping when calling them and so a `list[str]` could be passed which would simply be returned.

`ValueError` is raised if `groupby` parameter is not a list (i.e. string) and is `None` or not one of the currently four
supported values.