ci: Adds workflow to publish to PyPI; pre-commit gitleaks
Adds a workflow for publishing to PyPI. This is based on tagging commits with the pattern
`v*`. [PEP440](https://peps.python.org/pep-0440/) is the original specification for this, although as noted the current up-to-date specification can be found on the [Python Packaging Authority (PyPA)
Specifications](https://packaging.python.org/en/latest/specifications/version-specifiers).

> The canonical public version identifiers MUST comply with the following scheme:
>
>     [N!]N(.N)*[{a|b|rc}N][.postN][.devN]
>
> Public version identifiers MUST NOT include leading or trailing whitespace.
>
> Public version identifiers MUST be unique within a given distribution.
> ...
>
> Public version identifier are separated into up to five segments:
>
> - Epoch segment: N!
> - Release segment: N(.N)*
> - Pre-release segment: {a|b|rc}N
> - Post-release segment: .postN
> - Development release segment: .devN

Currently I would class IsoSLAM as pre-release and so we should start with `v0.0.1-a1` and increment `a#`.

As the process of making releases involves generating tokens which are to be stored in `.pypirc` (should manual release
be made/required), I have added a [gitleaks pre-commit hook](https://github.com/gitleaks/gitleaks#pre-commit) that checks these have not been included in commits as well as the traditional method of adding them to `.gitignore`.